task1:counter app
