+++ lib/task1/counter_screen.dart
@@ -1,0 +1,40 @@
+import 'package:flutter/material.dart';
+
+class CounterScreen extends StatefulWidget {
+  const CounterScreen({super.key});
+
+  @override
+  State<CounterScreen> createState() => _CounterScreenState();
+}
+
+class _CounterScreenState extends State<CounterScreen> {
+  int _counter = 0;
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Text(
+            _counter.toString(),
+            style: Theme.of(context).textTheme.displaySmall,
+          ),
+          IconButton(
+            onPressed: _incrementCount,
+            icon: Icon(
+              Icons.add,
+              color: Theme.of(context).colorScheme.primary,
+            ),
+          )
+        ],
+      ),
+    );
+  }
+
+  void _incrementCount() {
+    setState(() {
+      _counter++;
+    });
+  }
+}
