+++ test/widget_test.dart
@@ -6,14 +6,13 @@
 // tree, read text, and verify that the values of widget properties are correct.
 
 import 'package:flutter/material.dart';
+import 'package:flutter_study/task1/task1_app.dart';
 import 'package:flutter_test/flutter_test.dart';
-
-import 'package:flutter_study/main.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const MyApp());
+    await tester.pumpWidget(const Task1App());
 
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);
